Managed to manually authorize my application routes using Policies but I have not been currently able to Delete an attendee. *Bug

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,15 +8,19 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class AttendeeController extends Controller
 {
-    use CanLoadRelationships;
+    use CanLoadRelationships, AuthorizesRequests;
 
     private array $relations = ['user'];
 
-    public function index(Event $event)
+    public function index(Request $request, Event $event)
     {
+        // Authorize the viewAny action
+        $this->authorize('viewAny', [Attendee::class, $request->user()]);
+
         $attendees = $this->loadRelationships($event->attendees()->latest());
 
         return AttendeeResource::collection($attendees->paginate());
@@ -24,9 +28,12 @@ class AttendeeController extends Controller
 
     public function store(Request $request, Event $event)
     {
+        // Authorize the create action
+        $this->authorize('create', [Attendee::class, $request->user()]);
+
         $attendee = $this->loadRelationships(
             $event->attendees()->create([
-            'user_id' => 1
+                'user_id' => $request->user()->id
             ])
         );
 
@@ -35,11 +42,17 @@ class AttendeeController extends Controller
 
     public function show(Event $event, Attendee $attendee)
     {
+        // Authorize the view action
+        $this->authorize('view', $attendee->event);
+
         return new AttendeeResource($this->loadRelationships($attendee));
     }
 
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Request $request, Event $event, Attendee $attendee)
     {
+        // Authorize the update action
+        $this->authorize('update', [$event, $request->user()]);
+
         $attendee->delete();
 
         return response(status: 204);
